rates_app/rates_shared/requirements.txt
artifacts-keyring==0.3.1
astroid==2.5.2
autopep8==1.5.6
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coverage==5.5
docutils==0.17
flake8==3.9.0
idna==2.10
importlib-metadata==3.10.0
isort==5.8.0
keyring==23.0.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
packaging==20.9
pkginfo==1.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.8.1
pylint==2.7.4
pyparsing==2.4.7
pywin32-ctypes==0.2.0
PyYAML==6.0
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
toml==0.10.2
tqdm==4.59.0
twine==3.4.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
webencodings==0.5.1
wrapt==1.12.1
zipp==3.4.1
